pyproject.toml
[project]
name = "iLoveWord"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "NingmengLemon", email = "60469383+NingmengLemon@users.noreply.github.com"},
]
dependencies = [
    "pillow>=10.4.0",
    "paddleocr>=2.8.1",
    "fuzzywuzzy>=0.18.0",
    "uiautomator2>=3.2.5",
    "paddlepaddle>=2.6.2",
    "setuptools>=75.1.0",
]
requires-python = "==3.10.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
